WEB-3103    CLONE - [6.0.2] FileWidget in form generate 'File not found' and 'Error while getting the form page list' and 'NullPointerException'

+++ forms/forms-server/src/main/java/org/bonitasoft/forms/server/api/impl/FormExpressionsAPIImpl.java
@@ -275,6 +275,7 @@ public class FormExpressionsAPIImpl implements IFormExpressionsAPI {
      *            the field ID
      * @param fieldValues
      *            the form field values
+	 * @param deleteDocument
      * @return the value of the field as an Serializable
      * @throws FileTooBigException
      * @throws IOException
@@ -282,14 +283,14 @@ public class FormExpressionsAPIImpl implements IFormExpressionsAPI {
      * @throws InvalidSessionException
      */
     protected Serializable evaluateFieldValueActionExpression(final APISession session, final String fieldId, final Map<String, FormFieldValue> fieldValues,
-            final Locale locale) throws FileTooBigException, IOException, InvalidSessionException, BPMEngineException {
+            final Locale locale, final boolean deleteDocuments) throws FileTooBigException, IOException, InvalidSessionException, BPMEngineException {
 
         Serializable result = null;
 
         final FormFieldValue fieldValue = fieldValues.get(fieldId);
         if (fieldValue != null) {
             if (fieldValue.isDocument()) {
-                result = getDocumentValue(session, fieldValue);
+                result = getDocumentValue(session, fieldValue, deleteDocuments);
             } else {
                 result = fieldValue.getValue();
             }
@@ -302,14 +303,15 @@ public class FormExpressionsAPIImpl implements IFormExpressionsAPI {
      * 
      * @param session
      * @param fieldValue
+     * @param deleteDocument
      * @return
      * @throws FileTooBigException
      * @throws IOException
      * @throws BPMEngineException
      * @throws InvalidSessionException
      */
-    protected DocumentValue getDocumentValue(final APISession session, final FormFieldValue fieldValue) throws FileTooBigException, IOException,
-            InvalidSessionException, BPMEngineException {
+    protected DocumentValue getDocumentValue(final APISession session, final FormFieldValue fieldValue, final boolean deleteDocument)
+            throws FileTooBigException, IOException, InvalidSessionException, BPMEngineException {
         DocumentValue documentValue = null;
         final String uri = (String) fieldValue.getValue();
         if (File.class.getName().equals(fieldValue.getValueType())) {
@@ -329,7 +331,9 @@ public class FormExpressionsAPIImpl implements IFormExpressionsAPI {
                     final FileTypeMap mimetypesFileTypeMap = new MimetypesFileTypeMap();
                     final String contentType = mimetypesFileTypeMap.getContentType(theSourceFile);
                     documentValue = new DocumentValue(fileContent, contentType, originalFileName);
-                    theSourceFile.delete();
+                    if (deleteDocument) {
+                        theSourceFile.delete();
+                    }
                 } else {
                     if (LOGGER.isLoggable(Level.SEVERE)) {
                         LOGGER.log(Level.SEVERE, "Error while retrieving the uploaded file " + uri + ": File not found.");
@@ -365,6 +369,7 @@ public class FormExpressionsAPIImpl implements IFormExpressionsAPI {
      * @param session
      * @param fieldValues
      * @param locale
+     * @param deleteDocuments
      * @return the context
      * @throws IOException
      * @throws FileTooBigException
@@ -373,13 +378,14 @@ public class FormExpressionsAPIImpl implements IFormExpressionsAPI {
      */
     @Override
     public Map<String, Serializable> generateGroovyContext(final APISession session, final Map<String, FormFieldValue> fieldValues, final Locale locale,
-            Map<String, Serializable> context) throws FileTooBigException, IOException, InvalidSessionException, BPMEngineException {
+            Map<String, Serializable> context, final boolean deleteDocuments) throws FileTooBigException, IOException, InvalidSessionException,
+            BPMEngineException {
         if (context == null) {
             context = new HashMap<String, Serializable>();
         }
         for (final Entry<String, FormFieldValue> fieldValuesEntry : fieldValues.entrySet()) {
             final String fieldId = fieldValuesEntry.getKey();
-            final Serializable fieldValue = evaluateFieldValueActionExpression(session, fieldId, fieldValues, locale);
+            final Serializable fieldValue = evaluateFieldValueActionExpression(session, fieldId, fieldValues, locale, deleteDocuments);
             context.put(FIELDID_PREFIX + fieldId, fieldValue);
         }
         context.put(IFormExpressionsAPI.USER_LOCALE, locale);
@@ -441,7 +447,7 @@ public class FormExpressionsAPIImpl implements IFormExpressionsAPI {
             throws BPMEngineException, InvalidSessionException, FileTooBigException, IOException, ExpressionEvaluationException {
         Serializable result = null;
         if (expression != null) {
-            final Map<String, Serializable> evalContext = generateGroovyContext(session, fieldValues, locale, context);
+            final Map<String, Serializable> evalContext = generateGroovyContext(session, fieldValues, locale, context, false);
             final Map<org.bonitasoft.engine.expression.Expression, Map<String, Serializable>> expressionWithContext = new HashMap<org.bonitasoft.engine.expression.Expression, Map<String, Serializable>>();
             final ExpressionAdapter expressionAdapter = new ExpressionAdapter();
             expressionWithContext.put(expressionAdapter.getEngineExpression(expression), evalContext);
@@ -517,7 +523,7 @@ public class FormExpressionsAPIImpl implements IFormExpressionsAPI {
             throws BPMEngineException, InvalidSessionException, FileTooBigException, IOException {
         Serializable result = null;
         if (expression != null) {
-            final Map<String, Serializable> evalContext = generateGroovyContext(session, fieldValues, locale, context);
+            final Map<String, Serializable> evalContext = generateGroovyContext(session, fieldValues, locale, context, false);
             final Map<org.bonitasoft.engine.expression.Expression, Map<String, Serializable>> expressions = new HashMap<org.bonitasoft.engine.expression.Expression, Map<String, Serializable>>();
             final ExpressionAdapter expressionAdapter = new ExpressionAdapter();
             expressions.put(expressionAdapter.getEngineExpression(expression), evalContext);
@@ -582,7 +588,7 @@ public class FormExpressionsAPIImpl implements IFormExpressionsAPI {
             InvalidSessionException, FileTooBigException, IOException {
         Serializable result = null;
         if (expression != null) {
-            final Map<String, Serializable> evalContext = generateGroovyContext(session, fieldValues, locale, context);
+            final Map<String, Serializable> evalContext = generateGroovyContext(session, fieldValues, locale, context, false);
             final Map<org.bonitasoft.engine.expression.Expression, Map<String, Serializable>> expressionWithContext = new HashMap<org.bonitasoft.engine.expression.Expression, Map<String, Serializable>>();
             final ExpressionAdapter expressionAdapter = new ExpressionAdapter();
             expressionWithContext.put(expressionAdapter.getEngineExpression(expression), evalContext);
@@ -858,7 +864,7 @@ public class FormExpressionsAPIImpl implements IFormExpressionsAPI {
         }
 
         if (!expressionsWithContext.isEmpty()) {
-            context = generateGroovyContext(session, fieldValues, locale, context);
+            context = generateGroovyContext(session, fieldValues, locale, context, false);
             Map<String, Serializable> evaluated;
 
             final ExpressionEvaluatorEngineClient engineClient = getExpressionEvaluator(session);
@@ -975,7 +981,7 @@ public class FormExpressionsAPIImpl implements IFormExpressionsAPI {
         }
 
         if (!expressionsWithContext.isEmpty()) {
-            context = generateGroovyContext(session, fieldValues, locale, context);
+            context = generateGroovyContext(session, fieldValues, locale, context, false);
             result.putAll(getProcessInstanceExpressionEvaluator(session).evaluate(getProcessInstanceAccessor(session, processInstanceId),
                     expressionsWithContext, !isCurrentValue));
         }
@@ -1070,7 +1076,7 @@ public class FormExpressionsAPIImpl implements IFormExpressionsAPI {
         }
 
         if (!expressionsWithContext.isEmpty()) {
-            context = generateGroovyContext(session, fieldValues, locale, context);
+            context = generateGroovyContext(session, fieldValues, locale, context, false);
             result.putAll(getExpressionEvaluator(session).evaluateExpressionsOnProcessDefinition(processDefinitionID, expressionsWithContext));
         }
         return result;
